# backend/src/docs/v1.api-spec.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: GetOK API
  description: >-
    This API allows the UMU webapp to access Ministry of Education oracle database
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
  contact:
    name: OBIEE UMU
servers:
  - url: /api/v1
    description: This Server
security:
  - bearerAuth: []
  - OpenID: []
paths:
  /checks/status:
    get:
      summary: Returns status of correspondent APIs
      operationId: checkStatus
      tags:
        - Checks
      responses:
        '200':
          description: Responds if the correspondent API endpoint(s) are healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointList'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/appConfig/{acronym}':
    get:
      summary: Returns WebADE JSON response for the specified acronym
      operationId: appConfigGetAcronym
      tags:
        - Application Configuration
      parameters:
        - name: acronym
          in: path
          description: Name of the project
          required: true
          schema:
            type: string
      responses:
        '200':
          description: WebADE JSON response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppConfig'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /appConfigs:
    get:
      summary: Returns WebADE JSON response for all defined acronyms
      operationId: appConfigGetAll
      tags:
        - Application Configuration
      parameters:
        - name: acronym
          in: path
          description: Name of the project
          required: true
          schema:
            type: string
      responses:
        '200':
          description: WebADE JSON response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppConfigList'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /appConfigForm:
    post:
      summary: Submit a proposed application
      operationId: appConfigFormPost
      tags:
        - Application Configuration
      requestBody:
        description: Form fields required to generate a minimal Application Configuration
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppConfigForm'
      responses:
        '200':
          description: WebADE JSON response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppConfigResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /validations/rules:
    post:
      summary: Returns status of correspondent APIs
      operationId: validateFields
      tags:
        - Validation
      responses:
        '200':
          description: Responds if the API endpoint is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointList'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OpenID:
      type: openIdConnect
      openIdConnectUrl: 'https://example.com/.well-known/openid-configuration'
  schemas:
    AppConfig:
      type: object
      properties:
        '@type':
          type: string
          example: 'http://webade.gov.bc.ca/applicationConfiguration'
        applicationAcronym:
          type: string
          example: MSSC
        custodianNumber:
          type: integer
          minimum: 0
        applicationName:
          type: string
          example: Application Name
        applicationDescription:
          type: string
          example: Application Description
        applicationObjectPrefix:
          type: string
          example: null
        enabledInd:
          type: boolean
          example: true
        distributeTypeCd:
          type: string
          example: null
        managementEnabledInd:
          type: boolean
          example: false
        applicationVersion:
          type: string
          example: null
        reportedWebadeVersion:
          type: string
          example: null
        actions:
          type: array
          items:
            type: object
          example: []
        roles:
          type: array
          items:
            type: object
          example: []
        wdePreferences:
          type: array
          items:
            type: object
          example: []
        applicationPreferences:
          type: array
          items:
            type: object
          example: []
        globalPreferences:
          type: array
          items:
            type: object
          example: []
        defaultUserPreferences:
          type: array
          items:
            type: object
          example: []
        profiles:
          type: array
          items:
            type: object
          example: []
        serviceClients:
          type: array
          items:
            type: object
          example: []
        groupAuthorizations:
          type: array
          items:
            type: object
          example: []
    AppConfigForm:
      type: object
      required:
        - applicationAcronym
        - applicationName
        - applicationDescription
        - commonServices
        - passwordPublicKey
        - webadeEnvironment
      properties:
        applicationAcronym:
          type: string
          example: MSSC
        applicationName:
          type: string
          example: Application Name
        applicationDescription:
          type: string
          example: Application Description
        commonServices:
          type: array
          items:
            type: string
            example: CMSG
          description: An array of applicationAcronyms
        passwordPublicKey:
          type: string
          description: A public key in PEM format without the headers
          example: MIGfMA0GCSqGS...
        webadeEnvironment:
          type: string
          description: Which WebADE environment to work on - one of INT, TEST or PROD
          example: INT
    AppConfigList:
      type: array
      items:
        $ref: '#/components/schemas/AppConfig'
      description: A list of Application Configurations
    AppConfigResponse:
      required:
        - webAdeResponse
        - generatedPassword
        - generatedServiceClient
      properties:
        webAdeResponse:
          $ref: '#/components/schemas/AppConfig'
        generatedPassword:
          type: string
          description: 'Contains the generated password, encrypted with the user-supplied public key'
          example: ABC123XYZ
        generatedServiceClient:
          type: string
          description: 'Contains the service client name that was written to WebADE'
          example: DOMO_SERVICE_CLIENT
    EndpointStatus:
      type: object
      required:
        - name
        - endpoint
        - healthCheck
        - authenticated
        - authorized
      properties:
        name:
          type: string
          description: The name of the endpoint
          example: 'WebADE API'
        endpoint:
          type: string
          description: Base endpoint URL
          example: 'https://example.com/v1/'
        healthCheck:
          type: boolean
          description: Is the endpoint reachable
        authenticated:
          type: boolean
          description: Are credentials valid to access endpoint
        authorized:
          type: boolean
          description: Do the credentials have the right permissions
    EndpointList:
      type: object
      required:
        - endpoints
      properties:
        endpoints:
          type: array
          items:
            $ref: '#/components/schemas/EndpointStatus'
          description: A list of Endpoint Statuses
    Error:
      required:
        - error
      properties:
        error:
          type: string
          description: contains a message about the failure
          example: Something errored
    Success:
      required:
        - message
      properties:
        message:
          type: string
          description: 'Contains message about the success, normally "Written successfully"'
          example: Written successfully
  responses:
    Forbidden:
      description: Lack required role to perform action
    NotFound:
      description: Not found
    UnauthorizedError:
description: Access token is missing or invalid